move shape differs into patches folder
