+++ projects/optic/src/commands/capture/patches/patches.ts
@@ -2,9 +2,10 @@ import { Operation as JsonOps } from 'fast-json-patch';
 import { ParseResult } from '../../../utils/spec-loaders';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
 import { checkOpenAPIVersion } from '@useoptic/openapi-io';
+import { OpenAPIV3 } from '@useoptic/openapi-utilities';
 
 import { ApiCoverageCounter } from '../coverage/api-coverage';
-import { OpenAPIV3, SpecPatch, SpecPatches } from '../../oas/specs';
+import { SpecPatch, SpecPatches } from '../../oas/specs';
 import { CapturedInteractions } from '../sources/captured-interactions';
 import { DocumentedInteraction, Operation } from '../../oas/operations';
 import { DocumentedBodies } from '../../oas/shapes';
