+++ projects/optic/src/commands/oas/shapes/diffs/index.ts
@@ -1,22 +1,7 @@
 import {
-  ShapeDiffTraverser,
   JsonSchemaKnownKeyword,
   SchemaCompilationError,
-} from './traverser';
-import { ShapeDiffResult, ShapeDiffResultKind } from './result';
-import { Body } from '../body';
-import { SchemaObject } from '../schema';
-import { Result } from 'ts-results';
+} from '../../../capture/patches/patchers/shapes/diff';
 
-export type { ShapeDiffResult };
-export { ShapeDiffResultKind };
 export { JsonSchemaKnownKeyword };
 export { SchemaCompilationError };
-
-export function diffBodyBySchema(
-  body: Body,
-  schema: SchemaObject
-): Result<IterableIterator<ShapeDiffResult>, SchemaCompilationError> {
-  let traverser = new ShapeDiffTraverser();
-  return traverser.traverse(body.value, schema).map(() => traverser.results());
-}
